Rename cache to module-cache
