+++ lib/elm-files.js
@@ -4,13 +4,13 @@ const glob = require('glob');
 const fs = require('fs-extra');
 const chalk = require('chalk');
 const Hash = require('./hash');
-const Cache = require('./cache');
 const Debug = require('./debug');
 const Benchmark = require('./benchmark');
 const Anonymize = require('./anonymize');
 const elmParser = require('./parse-elm');
 const AppState = require('./state');
 const OsHelpers = require('./os-helpers');
+const ModuleCache = require('./module-cache');
 const ErrorMessage = require('./error-message');
 
 const defaultGlob = '**/*.elm$';
@@ -166,14 +166,14 @@ async function readFile(
 
 function readAst(options, elmParserPath, source) {
   const hash = Hash.hash(source);
-  return Cache.readAstFromFSCache(options, hash).then(async (result) => {
+  return ModuleCache.readAstFromFSCache(options, hash).then(async (result) => {
     if (result) {
       return result;
     }
 
     const ast = await elmParser.parse(elmParserPath, source);
     if (ast !== null) {
-      Cache.cacheFile(options, hash, ast);
+      ModuleCache.cacheFile(options, hash, ast);
     }
 
     return ast;
