+++ lib/module-cache.js
