+++ lib/runner.js
@@ -1,11 +1,11 @@
 #!/usr/bin/env node
 
 const Hash = require('./hash');
-const Cache = require('./cache');
 const Report = require('./report');
 const AppState = require('./state');
 const Autofix = require('./autofix');
 const AppWrapper = require('./app-wrapper');
+const ModuleCache = require('./module-cache');
 const ErrorMessage = require('./error-message');
 const {getProjectFiles} = require('./elm-files');
 const SuppressedErrors = require('./suppressed-errors');
@@ -53,7 +53,7 @@ function sendProjectContent(
       filesPendingReceiptAcknowledgement.delete(file.path);
 
       if (file.cacheRequest) {
-        await Cache.cacheFile(
+        await ModuleCache.cacheFile(
           options,
           Hash.hash(file.cacheRequest.source),
           file.cacheRequest.ast
@@ -97,7 +97,7 @@ async function initializeApp(options, elmModulePath, reviewElmJson, appHash) {
     process.exit(1);
   });
 
-  Cache.subscribe(options, app);
+  ModuleCache.subscribe(options, app);
   Autofix.subscribe(options, app);
   if (options.watch) {
     AppState.subscribe(app.ports.reviewReport, (result) => {
